Apply Clippy Fixes to Paging Code
This is better with all of the changes, but it is the Friday before
Christmas and I am trying to get this checked in and get out of
here. This won't break bisect, just an ugly commit.

+++ dxe_core/src/cpu_arch_protocol.rs
@@ -280,7 +280,7 @@ mod tests {
         let mut interrupt_manager = MockInterruptManager::new();
         interrupt_manager
             .expect_register_exception_handler()
-            .with(eq(ExceptionType::from(0 as usize)), always())
+            .with(eq(ExceptionType::from(0_usize)), always())
             .returning(|_, _| Ok(()));
         let protocol = EfiCpuArchProtocolImpl::new(&mut cpu_init, &mut interrupt_manager);
